feat: add movie details functionality
