+++ .eslintrc
@@ -1,5 +1,5 @@
 {
-  "extends": ["airbnb", "airbnb-typescript"],
+  "extends": ["airbnb", "airbnb-typescript", "plugin:react-hooks/recommended"],
   "parserOptions": {
     "warnOnUnsupportedTypeScriptVersion": false,
     "project": "./tsconfig.json"
@@ -11,6 +11,7 @@
     "react/destructuring-assignment": "off",
     "react/jsx-props-no-spreading": "off",
     "react/require-default-props": "off",
+    "jsx-a11y/control-has-associated-label": "off",
     "jsx-a11y/label-has-associated-control": ["error", {
       "required": {
         "some": ["nesting", "id"]
