+++ src/components/MovieContext/index.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export type TMovieContext = [number, React.Dispatch<React.SetStateAction<number>>];
+
+const MovieContext = createContext<TMovieContext | null>(null);
+
+export default MovieContext;
