+++ src/components/MovieDetails/_movieDetails.scss
@@ -1,0 +1,85 @@
+@import '../../scss/variables';
+
+.movieDetails {
+  border-bottom: 5px solid $light-grey;
+  padding: 20px 50px 30px;
+
+  &__header {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 30px;
+  }
+
+  &__searchButton {
+    background: url(../../assets/icons/search.svg) no-repeat;
+    background-size: contain;
+    width: 28px;
+    height: 29px;
+    cursor: pointer;
+  }
+
+  &__main {
+    display: flex;
+  }
+
+  &__poster {
+    margin-right: 60px;
+
+    img {
+      max-width: 323px;
+      height: 100%;
+    }
+  }
+
+  &__info {
+    &__title-container {
+      margin-bottom: 10px;
+    }
+
+    &__title {
+      font-size: 40px;
+      line-height: 49px;
+      letter-spacing: 1px;
+      font-weight: 300;
+      text-transform: uppercase;
+      margin-right: 25px;
+    }
+
+    &__rating {
+      font-weight: 300;
+      font-size: 20px;
+      line-height: 24px;
+      padding: 10px;
+      border: 1px solid $white;
+      border-radius: 50%;
+      vertical-align: super;
+    }
+
+    &__genre {
+      font-weight: 500;
+      font-size: 14px;
+      line-height: 17px;
+      opacity: .5;
+      margin-bottom: 30px;
+    }
+
+    &__additional {
+      color: $red;
+      font-size: 24px;
+      line-height: 29px;
+      font-weight: 300;
+      margin-bottom: 30px;
+
+      span:first-child {
+        margin-right: 50px;
+      }
+    }
+
+    &__overview {
+      font-weight: 300;
+      font-size: 20px;
+      line-height: 24px;
+      opacity: .5;
+    }
+  }
+}
